copy rather than symlink
Some PHP scripts use __DIR__, which for a symlink is the 'linked'
location, which doesn't match the 'logical' location when we join
things together by symlinking.

Ideally we'd eventually update upstream so no in-tree configuration is
needed anymore and everything can be done via environment variables.
Until then this works fine :)

fixes #9